examples/03-rag/docs_qa/urls.py
from django.urls import path
from . import views

urlpatterns = [
    path("", views.index, name="index"),
    path("ask/", views.ask, name="ask"),
    path("upload/", views.upload, name="upload"),
]
