examples/03-rag/requirements.txt
djangosdk>=0.1.1
pgvector>=0.3
pypdf>=4.0
psycopg2-binary>=2.9
python-decouple>=3.8
